backend/requirements.txt
aiohttp==3.9.5
asyncio==3.4.3
asyncpg==0.29.0
fastapi==0.111.0
passlib==1.7.4
PyJWT==2.8.0
python-dotenv==1.0.1
SQLAlchemy==2.0.30
uvicorn==0.30.1
